🚸 (new): Improve UI layout with Section Views and remove calculate toolbar button
Changed the previous nested VStacks inside the Form View to Sections so that the UI layout looks more organized with a clear header text for each section. Also removed the calculate toolbar button and the calculateBedTime func and replaced it with a computed property - "result". The value of the computed property is then always displayed to users under a new Section View with a header "Your ideal bedtime".